rename: 게시물 데이터 타입 파일 이름 변경
